Setup.sh

+++ Setup.sh
@@ -1,22 +1,31 @@
 #!/bin/bash
 
 echo " "
-echo "This script configures your machine with these vim and tmux settings for .NET development"
+echo "This script configures your machine with Neovim and F# autocomplete settings for .NET development"
 echo " "
 
-sudo apt update 
-
-sudo apt install curl neovim tmux -y
+# Detect the OS
+if [ -f /etc/debian_version ]; then
+    # For Ubuntu
+    sudo apt update 
+    sudo apt install curl neovim tmux -y
+elif [ -f /etc/fedora-release ]; then
+    # For Fedora
+    sudo dnf install curl neovim tmux -y
+else
+    echo "Unsupported operating system. Please use Ubuntu or Fedora."
+    exit 1
+fi
 
 echo " "
-echo "Installing vim plug for neovim ..."
+echo "Installing vim plug for Neovim ..."
 echo " "
 
 sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
        https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
 
 echo " "
-echo "Installing fsharp autocomplete ..."
+echo "Installing F# autocomplete ..."
 echo " "
 
 dotnet tool install -g fsautocomplete
@@ -24,18 +33,28 @@ dotnet tool install -g fsautocomplete
 echo " "
 echo "Configuring .tmux.conf and .init.vim ..."
 echo " "
-echo "$(cat Sources/_tmux.conf)" |  sudo tee  ~/.tmux.conf
 
-# reload config file (change file location to your the tmux.conf you want to use)
-bind r source-file ~/.tmux.conf
+# Create the .tmux.conf file
+if [ -f Sources/_tmux.conf ]; then
+    cat Sources/_tmux.conf | sudo tee ~/.tmux.conf
+else
+    echo "Source file for .tmux.conf not found."
+fi
 
-mkdir ~/.config/nvim
+# Reload config file command (for Tmux, user needs to run this manually)
+echo "To reload the Tmux configuration, run: 'tmux source-file ~/.tmux.conf'"
+
+mkdir -p ~/.config/nvim
 
 echo " "
-echo "$(cat Sources/_init.vim)" | sudo tee ~/.config/nvim/init.vim
+if [ -f Sources/_init.vim ]; then
+    cat Sources/_init.vim | sudo tee ~/.config/nvim/init.vim
+else
+    echo "Source file for init.vim not found."
+fi
 echo " "
 
 echo " "
-echo "Your neovim and tmux has been successfully configured ..."
+echo "Your Neovim and Tmux have been successfully configured ..."
 echo " "
